text-to-text/pyproject.toml
[tool.poetry]
name = "text-to-text"
version = "0.1.0"
description = ""
authors = ["Kevin Krumbach <kevin@krumbachs.de>"]
license = "MIT"
readme = "README.md"
#packages = [{include = "text_to_text"}]
packages = []

[tool.poetry.dependencies]
python = "~3.11"
python-pptx = "^0.6.21"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.12.0"
cx-Freeze = "^6.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
